Add basic SAM4E driver. Add ipconfigETHERNET_DRIVER_ADDS_UDP_CHECKSUM, ipconfigETHERNET_DRIVER_ADDS_IP_CHECKSUM, ipconfigETHERNET_DRIVER_CHECKS_IP_CHECKSUM and ipconfigETHERNET_DRIVER_CHECKS_UDP_CHECKSUM definitions.

+++ FreeRTOS-Plus/Source/FreeRTOS-Plus-UDP/include/FreeRTOS_Sockets.h
@@ -114,10 +114,10 @@ struct freertos_sockaddr
 
 	#define FreeRTOS_inet_ntoa( ulIPAddress, pucBuffer )									\
 										sprintf( ( char * ) ( pucBuffer ), "%d.%d.%d.%d",	\
-											( ( ulIPAddress ) & 0xffUL ),					\
-											( ( ( ulIPAddress ) >> 8UL ) & 0xffUL ),		\
-											( ( ( ulIPAddress ) >> 16UL ) & 0xffUL ),		\
-											( ( ( ulIPAddress ) >> 24UL ) & 0xffUL ) )
+											( int ) ( ( ulIPAddress ) & 0xffUL ),			\
+											( int ) ( ( ( ulIPAddress ) >> 8UL ) & 0xffUL ),\
+											( int ) ( ( ( ulIPAddress ) >> 16UL ) & 0xffUL ),\
+											( int ) ( ( ( ulIPAddress ) >> 24UL ) & 0xffUL ) )
 
 #else /* ipconfigBYTE_ORDER */
